pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "risk-engine"
version = "0.1.0"
description = "Portfolio risk & stress testing engine"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "pandas>=2.0",
  "numpy>=1.24",
  "yfinance>=0.2",
  "matplotlib>=3.7"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
